Move report adaption logic to own file
